+++ libft/libft.h
@@ -55,7 +55,7 @@ char			*ft_strtrim(char const *s1, char const *set);
 void			ft_putchar_fd(char c, int fd);
 void			ft_putstr_fd(char *s, int fd);
 void			ft_putendl_fd(char *s, int fd);
-void			ft_putnbr_fd(int n, int fd);
+void			ft_putnbr_fd(long n, int fd);
 char			*ft_itoa(int n);
 char			*ft_strmapi(char const *s, char (*f) (unsigned int, char));
 char			**ft_split(char const *s, char c);
@@ -70,7 +70,7 @@ void			ft_lstiter(t_list *lst, void (*f)(void *));
 t_list			*ft_lstmap(t_list *lst, void *(*f)(void *)
 				, void (*del)(void *));
 void			ft_hexa(size_t number);
-void			ft_hexamayus(size_t number);
+void			ft_hexamayus(unsigned int number);
 void			ft_memorypointer(char *str);
 void			ft_unsigned_fd(int n, int fd);
